+++ spec/dummy/db/migrate/20190716130830_add_age_to_tree.rb
@@ -1,0 +1,5 @@
+class AddAgeToTree < ActiveRecord::Migration
+  def up
+    add_column :trees, :age, :integer
+  end
+end