+++ spec/dummy/db/migrate/20190716135241_add_type_to_user.rb
@@ -1,0 +1,5 @@
+class AddTypeToUser < ActiveRecord::Migration
+  def up
+    add_column :users, :title, :integer
+  end
+end