+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Changed
 - Technical - Makes the JWT lighter and consistent across lianas.
+- Filters - Support complex/generic conditions chaining. 🛡
+- Technical - Upgrade Ruby version to 2.3.4 for the CI.
 
 ## RELEASE 3.2.0 - 2019-07-22
 ### Added